% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{add_relative_time}
\alias{add_relative_time}
\title{Add relative time column to a dataframe, and diversity at the new time zero}
\usage{
add_relative_time(df, start_size, num_parameters)
}
\arguments{
\item{df}{dataframe with columns including "Generation", "NumCells" and "DriverEdgeDiversity"}

\item{start_size}{value of NumCells at which new_time should be set to zero}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}
}
\value{
the same dataframe with additional columns: "new_time" is the time elapsed since
NumCells = start_size; "div0" is DriverDiversity when NumCells = start_size; 
"rank_div0" is div0 rescaled to a rank between 0 and 1
}
\description{
Add relative time column to a dataframe, and diversity at the new time zero
}
\examples{
num_parameters = count_parameters(system.file("extdata", "", 
package = "demonanalysis", mustWork = TRUE))
comb_df <- combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
add_relative_time(comb_df, start_size = 100, num_parameters = num_parameters)
}
