% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{adjust_mig_rate}
\alias{adjust_mig_rate}
\title{Two-dimensional adjusted migration rate}
\usage{
adjust_mig_rate(m, two_dim)
}
\arguments{
\item{m}{migration rate per cell, relative to birth rate}

\item{two_dim}{whether growth is in two dimensions}
}
\value{
The migration rate.
}
\description{
Adjust migration rate for probability that migrating cell lands in an already-occupied deme
}
\details{
Adjustment for two-dimensional growth entails multiplying the migration rate by 
factors estimated from simulation results. First factor: the number of cells at the edge 
of the population, relative to the number expected if the population were a disc. 
Second factor: for cells at the edge, the average proportion of nearest neighbours that are empty.
}
\examples{
adjust_mig_rate(1, TRUE)
}
