% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{all_statuses}
\alias{all_statuses}
\title{Return the final line of every error log in a batch of simulations.}
\usage{
all_statuses(input_dir)
}
\arguments{
\item{input_dir}{base input directory name}
}
\value{
final line of each error log
}
\description{
Return the final line of every error log in a batch of simulations.
}
