% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{ancestry}
\alias{ancestry}
\title{Get the ancestry of a genotype}
\usage{
ancestry(edges)
}
\arguments{
\item{edges}{Dataframe comprising an adjacency matrix, with column names "Parent" and "Identity"}
}
\description{
Get the ancestry of a genotype
}
\examples{
library(readr)
phylo <- read_delim("data/driver_phylo.dat", "\\t")
phylo <- filter(phylo, CellsPerSample == -1)
edges <- get_edges(phylo)
ancestry(edges)
}
