% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{apply_combinations_parallel}
\alias{apply_combinations_parallel}
\title{Apply a function to every combination of some sequences (parallelized version)}
\usage{
apply_combinations_parallel(n_cores, vec, fn, ...)
}
\arguments{
\item{n_cores}{number of cores}

\item{vec}{vector of final values of the sequences (initial values are always zero)}

\item{fn}{function to apply to the values}

\item{...}{other arguments passed to fn}
}
\value{
result of applying fn to every combination of vec values
}
\description{
Apply a function to every combination of some sequences (parallelized version)
}
\details{
Adding the argument type="FORK" to makeCluster solves the problem of variables and function scope
(this option means that the function automatically knows all environment variables)
!!!! Notice that this only works on mac and linux, else we would need to use clusterExport and clusterEvalQ
something like: clusterExport(cl, list("all_output", "combine_dfs", "get_population_df", "fread"))
!!!! to test if needed!!!!
}
