% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{cor_pval}
\alias{cor_pval}
\title{Function to return the p.value of a two.sided test, using the spearman correlation coefficient for the test
This function is created to work with find_correlations}
\usage{
cor_pval(Var1, Var2)
}
\arguments{
\item{Var1}{first variable of interest}

\item{Var2}{second variable of interest}
}
\value{
p.value of a two.sided test, using the spearman correlation coefficient for the test.
pvalue are computed via the asymptotic t approximation (exact=FALSE).
return NA if no p.value has been computed, instead of NULL.
}
\description{
Function to return the p.value of a two.sided test, using the spearman correlation coefficient for the test
This function is created to work with find_correlations
}
