% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_summary-data.R
\docType{data}
\name{cor_summary}
\alias{cor_summary}
\title{Correlations between diversities and "outcome"}
\format{Dataframe with one row for each unique combination of parameter values, gap and start_size 
(i.e. it summarises over "seed").}
\usage{
data(cor_summary)
}
\description{
Correlations useful for forecasting analyses.
}
\keyword{datasets}
