% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{count_drivers}
\alias{count_drivers}
\title{Count the number of ancestors of a node in a phylogenetic tree}
\usage{
count_drivers(edges, node)
}
\arguments{
\item{edges}{dataframe comprising an adjacency matrix}

\item{node}{node for which to calculate the metric}
}
\value{
Number of ancestors
}
\description{
Count the number of ancestors of a node in a phylogenetic tree
}
\examples{
edges1 <- data.frame(Parent = c(0,1,1,2,2,3,3), Identity = 1:7, Population = c(2,10,5,10,20,10,3))
count_drivers(edges1, 6)
}
