% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{count_seeds}
\alias{count_seeds}
\title{Count the number of rows in a dataframe per parameter set.}
\usage{
count_seeds(data, num_parameters)
}
\arguments{
\item{data}{dataframe}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}
}
\value{
a dataframe listing number of rows per parameter set
}
\description{
Count the number of rows in a dataframe per parameter set.
}
\examples{
count_seeds(sum_df, 16)
}
