% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{create_plots_batch}
\alias{create_plots_batch}
\title{Create image files for every simulation in a batch}
\usage{
create_plots_batch(input_dir, output_dir = NA, type = "plot")
}
\arguments{
\item{input_dir}{base input directory name}

\item{output_dir}{folder in which to save the image files}

\item{type}{what type of images to create: "plot" or "chart" or c("plot", "chart")}
}
\value{
a set of image files
}
\description{
Create image files for every simulation in a batch
}
