% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-data.R
\docType{data}
\name{data}
\alias{data}
\title{Example dataframe combining output from many simulations}
\format{Dataframe with one row per time point for each simulation}
\usage{
data(data)
}
\description{
Population sizes and other metrics over time for each simulation.
}
\keyword{datasets}
