% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_wait_cor_summary-data.R
\docType{data}
\name{depth_wait_cor_summary}
\alias{depth_wait_cor_summary}
\title{Correlations between diversities and "waiting_time" for different biopsy protocols}
\format{Dataframe with one row for each unique combination of parameter values and start_size 
(i.e. it summarises over "seed").}
\usage{
data(depth_wait_cor_summary)
}
\description{
Correlations useful for forecasting analyses.
}
\keyword{datasets}
