% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate}
\alias{disp_rate}
\title{Expected disperal speed, allowing dispersal to occur before fixation}
\usage{
disp_rate(r1, r2, K, d, m, selection = "birth", symmetric = FALSE)
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{selection}{one of "birth", "death", "birth & migration"}

\item{symmetric}{whether migration occurs in both directions}
}
\value{
A number, corresponding to the speed.
}
\description{
Expected disperal speed, allowing dispersal to occur before fixation
}
\examples{
disp_rate(1, 1.1, 2, 1, 0.1, "birth")
disp_rate(1, 1.1, 2, 1, 0.1, "death")
}
