% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_fission}
\alias{disp_rate_fission}
\title{Dispersal speed for deme fission model}
\usage{
disp_rate_fission(r2, K, m, migration_type = 2, two_dim = TRUE)
}
\arguments{
\item{r2}{birth rate relative to reference birth rate}

\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{migration_type}{2 or 3 (see details)}

\item{two_dim}{whether to adjust for two-dimensional growth}
}
\value{
The dispersal speed.
}
\description{
Dispersal speed for deme fission model
}
\details{
Assumes that a deme cannot attempt fission unless its population size reaches K.

Meaning of migration_type:
2: fission, such that fission rate is correlated with birth rate, and death rate is uniform
3: fission, such that fission rate is independent of birth rate, and death rate is uniform
}
\examples{
disp_rate_fission(1, 2, 0.1)
disp_rate_fission(1, 2, 0.1, migration_type = 3)

\dontrun{
# find dispersal rate from simulation:
df <- combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
num_parameters <- count_parameters(system.file("extdata", "", 
package = "demonanalysis", mustWork = TRUE))
df <- add_columns(df, num_parameters)
median(df$RadiusGrowthRate[which(df$NumCells > 400)])

# predicted dispersal rate:
disp_rate_fission(1, 1, 1)
}
}
