% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_max}
\alias{disp_rate_max}
\title{Maximum limit on dispersal speed}
\usage{
disp_rate_max(K, d, m, r1, r2, selection = "birth", symmetric = FALSE,
  two_dim = FALSE)
}
\arguments{
\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{selection}{one of "birth", "death", "birth & migration"}

\item{symmetric}{whether migration occurs in both directions}

\item{two_dim}{whether to adjust for two-dimensional growth}
}
\value{
A number, corresponding to the speed.
}
\description{
Maximum limit on dispersal speed
}
\details{
Assumes that the time until fixation is negligible relative to the time until migration. 
Adjustment for two-dimensional growth entails multiplying the migration rate by 0.4, which is
a factor estimated from simulation results.
}
\examples{
disp_rate_max(2, 1, 0.1, 1, 1.1, "birth")
disp_rate_max(2, 1, 0.1, 1, 1.1, "death")
}
