% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_min}
\alias{disp_rate_min}
\title{Minimum limit on dispersal speed for migration model}
\usage{
disp_rate_min(r1, r2, K, m, migration_type = 0, symmetric = FALSE,
  two_dim = TRUE, d = NA)
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{migration_type}{0 or 1 (see details)}

\item{symmetric}{whether migration occurs in both directions}

\item{two_dim}{whether to adjust for two-dimensional growth}

\item{d}{migration distance relative to 1/sqrt(K); if NA then d is set to sqrt(K)}
}
\value{
The lower limit on the growth rate of the radius, measured in cells (not demes).
}
\description{
Minimum limit on dispersal speed for migration model
}
\details{
Assumes that a cell type cannot attempt to migrate until it reaches fixation.

Meaning of migration_type:
\itemize{
  \item 0: migration, such that a migration attempt occurs only following a birth event
  \item 1: migration, such that migration is independent of birth
  }
}
\examples{
disp_rate_min(1, 1.1, 2, 0.1, 0)
disp_rate_min(1, 1.1, 2, 0.1, 1)
}
