% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{driver_geno_div_df_batch}
\alias{driver_geno_div_df_batch}
\title{Create a data frame of genotype diversity versus time for a batch of simulations}
\usage{
driver_geno_div_df_batch(input_dir)
}
\arguments{
\item{input_dir}{base input directory name}
}
\value{
a data frame
}
\description{
Create a data frame of genotype diversity versus time for a batch of simulations
}
\examples{
driver_geno_div_df_batch(system.file("example_batch", "", 
package = "demonanalysis", mustWork = TRUE))
}
