% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{driver_geno_div_plot_batch}
\alias{driver_geno_div_plot_batch}
\title{Create a plot of genotype diversity versus time for a batch of simulations}
\usage{
driver_geno_div_plot_batch(input_dir, output_filename = NA,
  file_type = "png", output_dir = NA)
}
\arguments{
\item{input_dir}{base input directory name}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
either an image file or a plot displyed on screen
}
\description{
Create a plot of genotype diversity versus time for a batch of simulations
}
\details{
Lines are coloured by combinations of parameter values.
}
\examples{
driver_geno_div_plot_batch(system.file("example_batch", "", 
package = "demonanalysis", mustWork = TRUE))
}
