% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_matrix-data.R
\docType{data}
\name{driver_matrix}
\alias{driver_matrix}
\title{Example relatedness matrix}
\format{Square matrix}
\usage{
data(driver_matrix)
}
\description{
Example matrix containing hamming distances between genotypes (upper triangular half only).
}
\keyword{datasets}
