% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver_phylo-data.R
\docType{data}
\name{driver_phylo}
\alias{driver_phylo}
\title{Example phylogeny dataframe}
\format{A dataframe with column names including "Generation", "Identity", "Parent", "Population", "CellsPerSample".}
\usage{
data(driver_phylo)
}
\description{
Example dataframe containing both phylogenetic information and population dynamics.
}
\keyword{datasets}
