% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{filter_by_generation_or_numcells}
\alias{filter_by_generation_or_numcells}
\title{Filter a data frame by Generation or NumCells}
\usage{
filter_by_generation_or_numcells(df, path, generation = NA, numcells = NA)
}
\arguments{
\item{df}{data frame}

\item{path}{folder containing output.dat (needed for numcells only)}

\item{generation}{Generation at which to filter (default NA corresponds to no filtering)}

\item{numcells}{Number of cells at which to filter (takes precedent over generation; default NA corresponds to no filtering)}
}
\value{
the combined dataframe
}
\description{
Filter a data frame by Generation or NumCells
}
\examples{
df <- read_delim_special(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE))
filter_by_generation_or_numcells(df, NA, 10, NA)
my_path <- system.file("extdata", package = "demonanalysis", mustWork = TRUE)
filter_by_generation_or_numcells(df, my_path, NA, 100)
}
