% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{final_error_message}
\alias{final_error_message}
\title{Return the final line of an error log of a simulation}
\usage{
final_error_message(input_dir)
}
\arguments{
\item{input_dir}{directory name}
}
\value{
final line of error log
}
\description{
Return the final line of an error log of a simulation
}
\examples{
final_error_message("data")
}
