% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{final_generation}
\alias{final_generation}
\title{Return the final generation number of a simulation}
\usage{
final_generation(input_dir)
}
\arguments{
\item{input_dir}{directory name}
}
\value{
generation number
}
\description{
Return the final generation number of a simulation
}
\examples{
final_generation(system.file("extdata", "", package = "demonanalysis"))
}
