% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{first_inc_moment}
\alias{first_inc_moment}
\title{Get the first incomplete moment from frequency data}
\usage{
first_inc_moment(sizes, counts, threshold)
}
\arguments{
\item{sizes}{for example, midpoints of a histogram}

\item{counts}{counts corresponding to the sizes}

\item{threshold}{lower bound of sizes to include}
}
\value{
the first incomplete moment
}
\description{
Get the first incomplete moment from frequency data
}
\examples{
df_test <- data.frame(size = 1:20, count = exp(-(1:20)))
first_inc_moment(df_test$size, df_test$count, 2)
}
