% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{get_common_allele_hist}
\alias{get_common_allele_hist}
\title{Plot a histogram of variant allele frequencies for the most common alleles}
\usage{
get_common_allele_hist(file, generation = NA)
}
\arguments{
\item{file}{name of file containing lists of allele frequencies}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
histogram object
}
\description{
Plot a histogram of variant allele frequencies for the most common alleles
}
\examples{
hist1 <- get_common_allele_hist(system.file("extdata", "output_allele_freqs.dat", 
package = "demonanalysis", mustWork = TRUE))
plot(hist1, xlim = c(0.1, 1), main = "", xlab = "allele frequency", ylab = "count")
}
