% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{get_genotype_sizes_hist}
\alias{get_genotype_sizes_hist}
\title{Get a histogram of genotype sizes}
\usage{
get_genotype_sizes_hist(file, generation = NA)
}
\arguments{
\item{file}{name of file containing lists of genotype sizes}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
histogram object
}
\description{
Get a histogram of genotype sizes
}
\examples{
hist1 <- get_genotype_sizes_hist(system.file("extdata", "genotypes.dat", 
package = "demonanalysis", mustWork = TRUE))
plot_genotype_sizes_hist(hist1, xmax = 50)
}
