% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{get_tree_from_matrix}
\alias{get_tree_from_matrix}
\title{Get a "phylo" object from a relatedness matrix}
\usage{
get_tree_from_matrix(matrix)
}
\arguments{
\item{matrix}{square matrix}
}
\description{
Get a "phylo" object from a relatedness matrix
}
\examples{
library(readr)
matrix <- read.delim("data/driver_matrix.dat", header=FALSE)
tree <- get_tree_from_matrix(matrix)
plot(tree, show.tip.label = FALSE)
}
