% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{get_wait_cor_summary}
\alias{get_wait_cor_summary}
\title{Generate summary dataframe of correlations with "waiting_time"}
\usage{
get_wait_cor_summary(summary, col_names_list, num_parameters, min_count)
}
\arguments{
\item{summary}{dataframe including columns named "seed", "Generation", "start_time", "start_size", "gap" and "waiting_time"}

\item{col_names_list}{char vector of column names in the summary dataframe}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}

\item{min_count}{minimum number of items in each column (otherwise result will be NA)}
}
\value{
Dataframe with one row for each unique combination of parameter values and start_size 
(i.e. it summarises over "seed"), and including columns containing the correlations between "waiting_time" 
and each variable in col_names_list.
}
\description{
Generate summary dataframe of correlations with "waiting_time"
}
\examples{
wait_cor_summary <- get_wait_cor_summary(sum_df, 
c("DriverDiversity", "DriverEdgeDiversity"), 16, min_count = 5)
depth_wait_cor_summary <- get_wait_cor_summary(sum_df, 
c(paste0("DriverDiversityFrom1SamplesAtDepth", 0:10), 
paste0("DriverDiversityFrom4SamplesAtDepth", 0:10)), 
16, min_count = 5)
}
