% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{grid_plot}
\alias{grid_plot}
\title{Plot a grid from a properly formatted data frame.}
\usage{
grid_plot(image_df, palette, discrete = FALSE, add_legend = FALSE,
  legend_title = "")
}
\arguments{
\item{image_df}{data frame formatted by image_df_from_grid_file}

\item{palette}{colour palette}

\item{discrete}{whether to use a discrete or continuous colour scale}

\item{add_legend}{whether to add a legend}

\item{legend_title}{text for legend title}
}
\value{
a plot object
}
\description{
Plot a grid from a properly formatted data frame.
}
\examples{
image_df <- image_df_from_grid_file("data/output_passengersgrid.dat")
grid_plot(image_df)
}
