% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{image_df_from_grid_file}
\alias{image_df_from_grid_file}
\title{Read a file containing grid states and process it into a dataframe for plotting.}
\usage{
image_df_from_grid_file(file, trim = -1)
}
\arguments{
\item{file}{file name including path}

\item{trim}{how many rows and columns to remove; if trim < 0 (default) then all rows and columns containing NA are removed}
}
\value{
a dataframe formatted for plotting
}
\description{
Read a file containing grid states and process it into a dataframe for plotting.
}
\examples{
image_df <- image_df_from_grid_file("data/output_passengersgrid.dat")
}
