% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{inv_Simpson_index}
\alias{inv_Simpson_index}
\title{Calculate the inverse Simpson index}
\usage{
inv_Simpson_index(pops)
}
\arguments{
\item{pops}{vector of population sizes (or frequencies)}
}
\value{
The inverse Simpson index (effective number of species)
}
\description{
Calculate the inverse Simpson index
}
\examples{
inv_Simpson_index(c(rep(1, 100)))
}
