% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{lambda_invasion}
\alias{lambda_invasion}
\title{Rate of migration followed by survival}
\usage{
lambda_invasion(i, K, d, m, r1, r2, selection = "birth")
}
\arguments{
\item{i}{number of cells of the focus type}

\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{selection}{one of "birth", "death", "birth & migration"}
}
\value{
A number, corresponding to the rate.
}
\description{
Rate of migration followed by survival
}
\examples{
lambda_invasion(1, 2, 1, 0.1, 1, 1.1, "birth")
lambda_invasion(1, 2, 1, 0.1, 1, 1.1, "death")
}
