% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{make_dir}
\alias{make_dir}
\title{Create a directory name including parameter names and values}
\usage{
make_dir(input_dir, pars, indices)
}
\arguments{
\item{input_dir}{base directory}

\item{pars}{vector of parameter names}

\item{indices}{vector of parameter values, of same length as pars}
}
\value{
directory name
}
\description{
Create a directory name including parameter names and values
}
\examples{
make_dir("my_dir", c("migration_edge_only", "mu_driver_birth", "seed"), rep(0,3))
}
