% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{make_image_file_name}
\alias{make_image_file_name}
\title{Create an image file name including parameter names and values}
\usage{
make_image_file_name(prefix, pars, indices)
}
\arguments{
\item{prefix}{prefix for file name}

\item{pars}{vector of parameter names}

\item{indices}{vector of parameter values, of same length as pars}
}
\value{
file name
}
\description{
Create an image file name including parameter names and values
}
\examples{
make_image_file_name("plot", c("migration_edge_only", "mu_driver_birth", "seed"), rep(0,3))
}
