% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{metrics}
\alias{metrics}
\title{Calculate average number of ancestors per node (n) and diversity (D) for a phylogenetic tree}
\usage{
metrics(data)
}
\arguments{
\item{edges}{dataframe comprising an adjacency matrix, including Population}
}
\value{
Vector containing average number of ancestors per node (n) and diversity (D)
}
\description{
Calculate average number of ancestors per node (n) and diversity (D) for a phylogenetic tree
}
\examples{
edges1 <- data.frame(Parent = c(0,1,1,2,2,3,3), Identity = 1:7, Population = c(2,10,5,10,20,10,3))
count_drivers(edges1, 6)
}
