% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{mig_rate}
\alias{mig_rate}
\title{Quadratic formula used to set migration rate in demon.cpp}
\usage{
mig_rate(K, edge_only = 0)
}
\arguments{
\item{K}{deme carrying capacity}

\item{edge_only}{whether migration occurs at the edge only}
}
\value{
A number, corresponding to the speed.
}
\description{
Quadratic formula used to set migration rate in demon.cpp
}
\details{
Assumes migration_type = 0, migration_edge_only = 0, migration_rate_scales_with_K = 1.
}
\examples{
mig_rate(8)
}
