% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{muller_df_from_file}
\alias{muller_df_from_file}
\title{Read a "phylo" dataframe and process it for plotting}
\usage{
muller_df_from_file(file)
}
\arguments{
\item{file}{file name including path}
}
\value{
a dataframe formatted for plotting
}
\description{
Read a "phylo" dataframe and process it for plotting
}
\examples{
Muller_df <- muller_df_from_file(system.file("extdata", 
"driver_phylo.dat", package = "demonanalysis", mustWork = TRUE))
}
