% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output-data.R
\docType{data}
\name{output}
\alias{output}
\title{Example output dataframe}
\format{A dataframe with column names including "Generation" and "Population".}
\usage{
data(output)
}
\description{
Example dataframe containing population metrics over time.
}
\keyword{datasets}
