% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_allele_counts-data.R
\docType{data}
\name{output_allele_counts}
\alias{output_allele_counts}
\title{Example allele counts dataframe}
\format{A dataframe with column names including "Generation", "Size", "Frequency", "Count".}
\usage{
data(output_allele_counts)
}
\description{
Example dataframe containing allele frequencies and counts.
}
\keyword{datasets}
