% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_allele_cum_dist.R
\docType{data}
\name{output_allele_cum_dist}
\alias{output_allele_cum_dist}
\title{Example cumulative counts of alleles, versus inverse frequency}
\format{Dataframe with columns "inverse_frequency" and "cumulative_count"}
\usage{
data(output_allele_cum_dist)
}
\description{
Cumulative count of mutations across a range of inverse frequencies.
}
\keyword{datasets}
