% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_allele_hist-data.R
\docType{data}
\name{output_allele_hist}
\alias{output_allele_hist}
\title{Example allele frequency histogram data}
\format{Dataframe with columns "frequency" and "density"}
\usage{
data(output_allele_hist)
}
\description{
Density of mutations across a range of frequencies.
}
\keyword{datasets}
