% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_genotype_counts-data.R
\docType{data}
\name{output_genotype_counts}
\alias{output_genotype_counts}
\title{Example genotype counts dataframe}
\format{A dataframe with column names including "Generation", "Size", "Frequency", "Count".}
\usage{
data(output_genotype_counts)
}
\description{
Example dataframe containing genotype frequencies and counts.
}
\keyword{datasets}
