% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_passengersgrid-data.R
\docType{data}
\name{output_passengersgrid}
\alias{output_passengersgrid}
\title{Example grid}
\format{Square matrix.}
\usage{
data(output_passengersgrid)
}
\description{
Example matrix containing spatial data.
}
\keyword{datasets}
