% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{parameter_names_and_values}
\alias{parameter_names_and_values}
\title{Find the names and final values of parameters that were varied in a batch of simulations}
\usage{
parameter_names_and_values(input_dir)
}
\arguments{
\item{input_dir}{base input directory name}
}
\value{
dataframe of parameter names and values
}
\description{
Find the names and final values of parameters that were varied in a batch of simulations
}
