% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters-data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Example parameters file}
\format{Dataframe with one row containing parameter values, and column names corresponding to parameter names.}
\usage{
data(parameters)
}
\description{
Values of simulation parameters.
}
\keyword{datasets}
