% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_allele_hist}
\alias{plot_allele_hist}
\title{Plot a histogram of variant allele frequencies with logit x-axis and log y-axis}
\usage{
plot_allele_hist(file, generation = NA)
}
\arguments{
\item{file}{file containing columns "Frequency" and "Count"}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
plot displyed on screen
}
\description{
Plot a histogram of variant allele frequencies with logit x-axis and log y-axis
}
\examples{
plot_allele_hist(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE))
}
