% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_common_allele_counts}
\alias{plot_common_allele_counts}
\title{Plot counts of variant allele frequencies on linear scales}
\usage{
plot_common_allele_counts(file, generation = NA)
}
\arguments{
\item{file}{file containing columns "Frequency" and "Count"}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
plot displyed on screen
}
\description{
Plot counts of variant allele frequencies on linear scales
}
\examples{
plot_common_allele_counts(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE))
}
