% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_common_allele_hist}
\alias{plot_common_allele_hist}
\title{Plot a histogram of variant allele frequencies for the most common alleles}
\usage{
plot_common_allele_hist(file, generation = NA)
}
\arguments{
\item{file}{name of file containing lists of allele frequencies}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
histogram object
}
\description{
Plot a histogram of variant allele frequencies for the most common alleles
}
\examples{
plot_common_allele_hist(system.file("extdata", "output_allele_freqs.dat", 
package = "demonanalysis", mustWork = TRUE))
}
