% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_corr_outcome_versus_period}
\alias{plot_corr_outcome_versus_period}
\title{Plot correlation between outcome and diversity, versus projection period, coloured by K, for each value of tumour size at measurement}
\usage{
plot_corr_outcome_versus_period(df, output_filename = NA, file_type = "png",
  output_dir = NA)
}
\arguments{
\item{df}{dataframe generated by get_cor_summary}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot correlation between outcome and diversity, versus projection period, coloured by K, for each value of tumour size at measurement
}
\examples{
plot_corr_outcome_versus_period(cor_summary)
}
