% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_corr_waiting_time_versus_start_size}
\alias{plot_corr_waiting_time_versus_start_size}
\title{Plot correlation between waiting time and diversity, versus tumour size at measurement, coloured by K}
\usage{
plot_corr_waiting_time_versus_start_size(df,
  output_filename = "corr_waiting_time_versus_start_size",
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{dataframe generated by get_wait_cor_summary}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot correlation between waiting time and diversity, versus tumour size at measurement, coloured by K
}
\examples{
plot_corr_waiting_time_versus_start_size(wait_cor_summary)
}
