% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot counts of variant allele frequencies on linear scales}
\usage{
plot_counts(file_or_dataframe, generation = NA, ...)
}
\arguments{
\item{file_or_dataframe}{file or data frame containing columns "Frequency" and "Count"}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}

\item{...}{other parameters passed to plot}
}
\value{
plot displyed on screen
}
\description{
Plot counts of variant allele frequencies on linear scales
}
\examples{
plot_counts(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE), ylim = c(0, 10))
}
