% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_curves_faceted}
\alias{plot_curves_faceted}
\title{Plot relationship between two variables faceted by K (rows) and combinations of other parameters (columns)}
\usage{
plot_curves_faceted(df, num_parameters, x_var = "gen_adj",
  y_var = "NumCells", log = FALSE, output_filename = NA,
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{data frame}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}

\item{x_var}{column name of the x-variable (default "gen_adj")}

\item{y_var}{column name of the y-variable (default "NumCells")}

\item{log}{if TRUE then y-axis will be log-transformed (default FALSE)}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot relationship between two variables faceted by K (rows) and combinations of other parameters (columns)
}
\examples{
plot_curves_faceted(data, 16)
plot_curves_faceted(data, 16, x_var = "Generation", y_var = "MeanBirthRate")
}
