% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_driver_geno_div}
\alias{plot_driver_geno_div}
\title{Plot genotype diversity versus time}
\usage{
plot_driver_geno_div(sum_df, output_filename = NA, file_type = "png",
  output_dir = NA, log_y = FALSE, facet1 = "migration_type",
  facet2 = "migration_edge_only", height = 1, width = 1,
  ymax = 100, selected_seed = NA)
}
\arguments{
\item{sum_df}{data frame created by driver_geno_div_df_batch}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}

\item{log_y}{whether to apply a log10 transformation to the y axis (default FALSE)}

\item{facet1}{column name for facet rows}

\item{facet2}{column name for facet columns}

\item{height}{relative image height}

\item{width}{relative image width}

\item{ymax}{maximum of y axis}

\item{selected_seed}{seed number for which results will be highlighted}
}
\value{
either an image file or a plot displyed on screen
}
\description{
Plot genotype diversity versus time
}
\details{
Lines are coloured by combinations of parameter values.
}
\examples{
sum_df <- driver_geno_div_df_batch(system.file("example_batch", "", 
package = "demonanalysis", mustWork = TRUE))
plot_driver_geno_div(sum_df)
}
