% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_driver_genotype_freq_hist}
\alias{plot_driver_genotype_freq_hist}
\title{Plot a histogram of driver genotype frequencies for the most common driver genotypes}
\usage{
plot_driver_genotype_freq_hist(hist)
}
\arguments{
\item{hist}{histogram}
}
\value{
plot displyed on screen
}
\description{
Plot a histogram of driver genotype frequencies for the most common driver genotypes
}
\examples{
hist1 <- get_genotype_sizes_hist(system.file("extdata", "driver_genotypes.dat", 
package = "demonanalysis", mustWork = TRUE), xmax = NA, freqs = TRUE)
plot_driver_genotype_freq_hist(hist1)
}
