% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_figure2}
\alias{plot_figure2}
\title{Create a set of plots for Figure 2}
\usage{
plot_figure2(path, output_filename = NA, file_type = "png",
  output_dir = NA, trim = -1)
}
\description{
Create a set of plots for Figure 2
}
