% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_logit_freq_dist}
\alias{plot_logit_freq_dist}
\title{Plot a histogram of variant allele frequencies with logit x-axis and log y-axis}
\usage{
plot_logit_freq_dist(file_or_dataframe, generation = NA, ...)
}
\arguments{
\item{file_or_dataframe}{file or data frame containing columns "Frequency" and "Count"}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}

\item{...}{other parameters passed to plot}
}
\value{
plot displyed on screen
}
\description{
Plot a histogram of variant allele frequencies with logit x-axis and log y-axis
}
\examples{
plot_logit_freq_dist(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE))
}
