% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_mutation_waves}
\alias{plot_mutation_waves}
\title{Plot numers of cells with n mutations for each value of n}
\usage{
plot_mutation_waves(df)
}
\arguments{
\item{df}{data frame with columns "Generation", "CellsWith0Drivers" and "CellsWithXDrivers" for some values of X >= 1}
}
\value{
plot displyed on screen
}
\description{
Plot numers of cells with n mutations for each value of n
}
\examples{
plot_mutation_waves(output)
}
