% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_outcome_versus_waiting_time}
\alias{plot_outcome_versus_waiting_time}
\title{Plot outcome versus waiting time, coloured by K}
\usage{
plot_outcome_versus_waiting_time(df, output_filename = NA,
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{data frame}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot outcome versus waiting time, coloured by K
}
