% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_R2_verus_sweep_metric}
\alias{plot_R2_verus_sweep_metric}
\title{Plot R-squared versus mean sweep metric}
\usage{
plot_R2_verus_sweep_metric(df, output_filename = "R2_verus_sweep_metric",
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{dataframe generated by get_summary (filtered), get_cor_summary (filtered) or get_wait_cor_summary (filtered)}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot R-squared versus mean sweep metric
}
\examples{
plot_R2_verus_sweep_metric(wait_cor_summary)
}
