% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_scatter_waiting_time_diversity}
\alias{plot_scatter_waiting_time_diversity}
\title{Example scatter plots of waiting time versus diversity}
\usage{
plot_scatter_waiting_time_diversity(df, col_name = "DriverDiversity",
  output_filename = "scatter_waiting_time_diversity",
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{dataframe}

\item{col_name}{name of column containing a diversity metric (default "DriverDiversity")}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Example scatter plots of waiting time versus diversity
}
\examples{
library(dplyr)
plot_scatter_waiting_time_diversity(filter(summary, K == 16, gap == min(summary$gap)))
}
