% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_trajectories_by_diversity}
\alias{plot_trajectories_by_diversity}
\title{Plot growth trajectories, coloured by driver edge diversity}
\usage{
plot_trajectories_by_diversity(df,
  output_filename = "trajectories_by_diversity", file_type = "png",
  output_dir = NA)
}
\arguments{
\item{df}{data frame}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot growth trajectories, coloured by driver edge diversity
}
\examples{
plot_trajectories_by_diversity(data)
}
