% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_trajectories_faceted}
\alias{plot_trajectories_faceted}
\title{Plot growth trajectories faceted by K (rows) and combinations of other parameters (columns)}
\usage{
plot_trajectories_faceted(df, num_parameters, x_var = "gen_adj",
  output_filename = "trajectories_faceted", file_type = "png",
  output_dir = NA)
}
\arguments{
\item{df}{data frame}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}

\item{x_var}{column name of the x-variable (default "gen_adj")}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot growth trajectories faceted by K (rows) and combinations of other parameters (columns)
}
\examples{
plot_trajectories_faceted(data, 16)
plot_trajectories_faceted(data, 16, x_var = "Generation")
}
