% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot a phylogenetic tree}
\usage{
plot_tree(edges, output_dir = NA, output_filename = NA,
  fill = "black", display = TRUE)
}
\arguments{
\item{edges}{dataframe comprising an adjacency matrix (and optionally including Population column)}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}

\item{output_filename}{name of output image file}

\item{fill}{fill colour of the nodes (excluding the root, which is always red)}

\item{display}{whether to display the tree}
}
\value{
either an image file or a plot displyed on screen
}
\description{
Plot a phylogenetic tree
}
\examples{
edges1 <- data.frame(Parent = c(0,1,1,2,2,3,3), Identity = 1:7, Population = c(2,10,5,10,20,10,3))
plot_tree(edges1)
}
