% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_viol_sweep_metric}
\alias{plot_viol_sweep_metric}
\title{Violin plot of mean sweep metric (genotype frequency autocorrelation)}
\usage{
plot_viol_sweep_metric(df, output_filename = NA, file_type = "png",
  output_dir = NA)
}
\arguments{
\item{df}{dataframe generated by get_summary}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Violin plot of mean sweep metric (genotype frequency autocorrelation)
}
\examples{
plot_viol_sweep_metric(summary)
}
