% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{prob_successful_migration}
\alias{prob_successful_migration}
\title{For cells at the boundary of a population, find mean proportion of nearest neighbour sites that are empty}
\usage{
prob_successful_migration(mat)
}
\arguments{
\item{mat}{matrix in which empty sites are denoted NA}
}
\value{
a vector containing total number of empty neighbours, total number of edge cells, 
and average proportion of empty neighbours per edge cell
}
\description{
For cells at the boundary of a population, find mean proportion of nearest neighbour sites that are empty
}
\examples{
mat <- image_df_from_grid_file(system.file("extdata", 
"output_popgrid.dat", package = "demonanalysis", mustWork = TRUE), as_matrix = TRUE)
prob_successful_migration(mat)

# Note that the edge of the grid counts as empty space:
mat <- matrix(c(NA,NA,NA,NA,
1,1,1,1,
NA,NA,NA,NA), nrow = 3, byrow = TRUE)
prob_successful_migration(mat)
}
