% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{quadratic_diversity}
\alias{quadratic_diversity}
\title{Rao's quadratic diversity, using a binary distance measure, converted to the effective number of species}
\usage{
quadratic_diversity(value, freq, sigma, threshold = 0.1)
}
\arguments{
\item{value}{vector of trait values}

\item{freq}{vector of frequencies or abundances}

\item{sigma}{Cutoff for considering two trait values to be identical}

\item{threshold}{minimum trait value to include in calculation}
}
\value{
effective number of species
}
\description{
Rao's quadratic diversity, using a binary distance measure, converted to the effective number of species
}
\examples{
quadratic_diversity(1:5, 1:5, 1)
}
