% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{read_delim_special}
\alias{read_delim_special}
\title{Attempt to read a tab-delimited file and return the contents, or NA if the file doesn't exist}
\usage{
read_delim_special(file)
}
\arguments{
\item{file}{tab-delimited file}
}
\value{
the contents, or NA if the file doesn't exist
}
\description{
Attempt to read a tab-delimited file and return the contents, or NA if the file doesn't exist
}
\examples{
read_delim_special(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE))
}
