% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{reflect_on_diagonal}
\alias{reflect_on_diagonal}
\title{Reflect a matrix on its leading diagonal (useful if distance matrix has entries in upper half only)}
\usage{
reflect_on_diagonal(matrix)
}
\arguments{
\item{matrix}{square matrix}
}
\value{
square matrix of same dimensions as input
}
\description{
Reflect a matrix on its leading diagonal (useful if distance matrix has entries in upper half only)
}
\examples{
reflect_on_diagonal(matrix(1:9, ncol = 3))
}
