% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{refresh_data_files}
\alias{refresh_data_files}
\title{Don't use this; it's just for refreshing data files on Rob's machine}
\usage{
refresh_data_files()
}
\value{
NA
}
\description{
Don't use this; it's just for refreshing data files on Rob's machine
}
