% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{rho}
\alias{rho}
\title{Probability of fixation in a Moran process}
\usage{
rho(i, K, r)
}
\arguments{
\item{i}{number of cells of the focus type}

\item{K}{deme carrying capacity}

\item{r}{relative fitness of the focus type}
}
\value{
The fixation probability.
}
\description{
Probability of fixation in a Moran process
}
\examples{
rho(1, 8, 1.1)
}
