% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_df-data.R
\docType{data}
\name{sum_df}
\alias{sum_df}
\title{Example summary data}
\format{Dataframe}
\usage{
data(sum_df)
}
\description{
Dataframe containing summary metrics of multiple simulations.
}
\keyword{datasets}
