% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary-data.R
\docType{data}
\name{summary}
\alias{summary}
\title{Summary data for many simulations}
\format{Dataframe with one row per combination of gap and final_size for each simulation}
\usage{
data(summary)
}
\description{
Metrics useful for forecasting analyses.
}
\keyword{datasets}
