% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{sweep_sequence}
\alias{sweep_sequence}
\title{Calculate how the genotype frequency distribution changes over time}
\usage{
sweep_sequence(pop_df, lag_type = "generations", breaks = 10,
  lag_gens = 500)
}
\arguments{
\item{pop_df}{Dataframe with column names "Identity", "Population" and "Generation"}

\item{lag_type}{Either "generations" or "proportions" (default "generations")}

\item{breaks}{Number of breaks (used only if lag_type = "proportions"; default 10)}

\item{lag_gens}{Number of generations between  (used only if lag_type = "generations"; default 500)}
}
\description{
Calculate how the genotype frequency distribution changes over time
}
\examples{
library(readr)
library(ggmuller)
phylo <- read_delim("data/driver_phylo.dat", "\\t")
phylo <- filter(phylo, CellsPerSample == -1)
pop_df <- get_population_df(phylo)
sweep_seq1 <- sweep_sequence(pop_df, lag_type = "proportions", breaks = 6)
sweep_seq2 <- sweep_sequence(pop_df, lag_type = "generations", lag_gens = 2)
identical(sweep_seq1, sweep_seq2)
sweep_df <- data.frame(y = sweep_seq2, x = (1:length(sweep_seq2))/length(sweep_seq2))
plot(y ~ x , data = sweep_df, type = "l")
}
