% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{time_expected}
\alias{time_expected}
\title{Expected time of first successful migration, which may occur before fixation}
\usage{
time_expected(r1, r2, K, d, m, selection = "birth")
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{selection}{one of "birth", "death", "birth & migration"}
}
\value{
A number, corresponding to the time.
}
\description{
Expected time of first successful migration, which may occur before fixation
}
\examples{
time_expected(1, 1.1, 2, 1, 0.1, "birth")
time_expected(1, 1.1, 2, 1, 0.1, "death")
}
