% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{time_fixation}
\alias{time_fixation}
\title{Mean time until one cell achieves fixation, given that it reaches fixation}
\usage{
time_fixation(r1, r2, K, r_powers_shifted = NA)
}
\arguments{
\item{r1}{birth rate of non-focus cells, relative to reference birth rate}

\item{r2}{birth rate of focus cell type}

\item{K}{deme carrying capacity}

\item{r_powers_shifted}{optional look-up vector}
}
\value{
A number, corresponding to the time.
}
\description{
Mean time until one cell achieves fixation, given that it reaches fixation
}
\details{
See Traulsen 2009, equation 1.38.
}
\examples{
time_fixation(1, 1.1, 4)
}
