% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{time_migration}
\alias{time_migration}
\title{Migration waiting time from fully occupied deme}
\usage{
time_migration(K, m, r1, r2, migration_type, d = NA)
}
\arguments{
\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{migration_type}{0 or 1 (see details)}

\item{d}{migration distance relative to 1/sqrt(K); if NA then d is set to sqrt(K)}
}
\value{
The waiting time.
}
\description{
Mean time until successful migration from a deme that is fully occupied by the focus cell type
}
\details{
Meaning of migration_type:
\itemize{
  \item 0: migration, such that a migration attempt occurs only following a birth event
  \item 1: migration, such that migration is independent of birth
  }
}
\examples{
time_migration(2, 0.1, 1, 1.1, 0)
time_migration(2, 0.1, 1, 1.1, 1)
}
