% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wait_cor_summary-data.R
\docType{data}
\name{wait_cor_summary}
\alias{wait_cor_summary}
\title{Correlations between diversities and "waiting_time"}
\format{Dataframe with one row for each unique combination of parameter values and start_size 
(i.e. it summarises over "seed").}
\usage{
data(wait_cor_summary)
}
\description{
Correlations useful for forecasting analyses.
}
\keyword{datasets}
