% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{add_columns}
\alias{add_columns}
\title{Add derived variables to a dataframe}
\usage{
add_columns(df)
}
\arguments{
\item{df}{dataframe with columns including "Generation"}
}
\value{
the same dataframe with additional columns
}
\description{
Add derived variables to a dataframe
}
\examples{
add_columns(output)
}
