% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{adjust_mig_rate}
\alias{adjust_mig_rate}
\title{Two-dimensional adjusted migration rate}
\usage{
adjust_mig_rate(m, two_dim)
}
\arguments{
\item{m}{migration rate per cell, relative to birth rate}

\item{two_dim}{whether growth is in two dimensions}
}
\value{
The migration rate.
}
\description{
Adjust migration rate for probability that migrating cell lands in an already-occupied deme
}
\details{
Adjustment for two-dimensional growth entails multiplying the migration rate by 0.4, which is
a factor estimated from simulation results.
}
\examples{
adjust_mig_rate(1, TRUE)
}
