% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{all_output}
\alias{all_output}
\title{Create a composite dataframe for a batch of simulations, derived from multiple 
data files per simulation.}
\usage{
all_output(input_dir, include_diversities = TRUE)
}
\arguments{
\item{input_dir}{base input directory name}

\item{include_diversities}{boolean whether to include diversity metrics}
}
\value{
a combined dataframe
}
\description{
Create a composite dataframe for a batch of simulations, derived from multiple 
data files per simulation.
}
