% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{all_output}
\alias{all_output}
\title{Create a composite dataframe for a batch of simulations, derived from multiple 
data files per simulation.}
\usage{
all_output(input_dir, include_diversities = TRUE, df_type = "output",
  vaf_cut_off = NA)
}
\arguments{
\item{input_dir}{base input directory name}

\item{include_diversities}{boolean whether to include diversity metrics}

\item{df_type}{which dataframes to combine}

\item{vaf_cut_off}{exclude genotypes with vaf lower cut off from combined_df}
}
\value{
a combined dataframe
}
\description{
Create a composite dataframe for a batch of simulations, derived from multiple 
data files per simulation.
}
\examples{
all_output(system.file("example_batch", "", package = "demonanalysis", mustWork = TRUE))
all_output(system.file("example_batch", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "driver_genotype_properties")
all_output(system.file("example_batch", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "genotype_properties", vaf_cut_off = 0.002)
}
