% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{all_output}
\alias{all_output}
\title{Create a composite dataframe by combining data for every simulation in a batch.}
\usage{
all_output(input_dir, pars, final_values)
}
\arguments{
\item{input_dir}{base input directory name}

\item{pars}{vector of parameter names}

\item{final_values}{vector of largest parameter values, of same length as pars}
}
\value{
a combined dataframe
}
\description{
Create a composite dataframe by combining data for every simulation in a batch.
}
