% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{all_statuses}
\alias{all_statuses}
\title{Return the final line of every error log in a batch of simulations.}
\usage{
all_statuses(input_dir, pars, final_values)
}
\arguments{
\item{input_dir}{base input directory name}

\item{pars}{vector of parameter names}

\item{final_values}{vector of largest parameter values, of same length as pars}
}
\value{
final line of each error log
}
\description{
Return the final line of every error log in a batch of simulations.
}
