% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{all_statuses}
\alias{all_statuses}
\title{Return the final line (or an earlier line) of every error log in a batch of simulations.}
\usage{
all_statuses(input_dir, adjust = 0)
}
\arguments{
\item{input_dir}{base input directory name}

\item{adjust}{number of lines prior to the final line (default 0)}
}
\value{
one line from each error log
}
\description{
Return the final line (or an earlier line) of every error log in a batch of simulations.
}
