% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{ancestry}
\alias{ancestry}
\title{Get the ancestry of a genotype}
\usage{
ancestry(edges)
}
\arguments{
\item{edges}{Dataframe comprising an adjacency matrix, with column names "Parent" and "Identity"}
}
\description{
Get the ancestry of a genotype
}
\examples{
library(dplyr)
phylo <- filter(driver_phylo, CellsPerSample == -1)
library(ggmuller)
edges <- get_edges(phylo)
ancestry(edges)
}
