% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{combine_dfs}
\alias{combine_dfs}
\title{Combine (and optionally process) data for one simulation}
\usage{
combine_dfs(full_dir, include_diversities = TRUE, df_type = "output",
  max_generation = FALSE, vaf_cut_off = NA, generation = NA,
  numcells = NA, num_parameters = NA, ExitCode4 = FALSE)
}
\arguments{
\item{full_dir}{base input directory name}

\item{include_diversities}{boolean whether to include diversity metrics (if df_type == "output")}

\item{df_type}{which dataframes to combine}

\item{max_generation}{If TRUE then results are returned only for the final generation}

\item{vaf_cut_off}{exclude genotypes with vaf lower than cut off (if df_type == "genotype_properties" or "driver_genotype_properties")}

\item{generation}{Generation at which to filter (default NA corresponds to no filtering)}

\item{numcells}{Number of cells at which to filter (default NA corresponds to no filtering)}

\item{num_parameters}{Number of parameters, accounting for the first set of columns in the dataframe (optional, but may improve speed)}

\item{ExitCode4}{Allow to deal with simulations including treatment. Simulations for which the tumor die at the treatment time end with "Exit Code =4",
we need to deal in a specific way (i.e need to remove the lines with Treated==1) with these simulations when combining the output
(as these simulations have most columns containing missing valeus/NA after the treatment).}
}
\value{
the combined dataframe
}
\description{
Combine (and optionally process) data for one simulation
}
\details{
If both \code{generation} and \code{numcells} are provided then \code{numcells} 
takes precedent.
}
\examples{
combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "driver_genotype_properties")
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "genotype_properties", vaf_cut_off = 0.002)
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "genotype_counts", numcells = 100)
}
