% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{combine_dfs}
\alias{combine_dfs}
\title{Combine data for one simulation, including population metrics, parameter values 
and diversity metrics, and with added columns containing derived variables.}
\usage{
combine_dfs(full_dir, res = data.frame())
}
\arguments{
\item{full_dir}{base input directory name}

\item{res}{dataframe to which the result will be appended (default is an empty dataframe)}
}
\value{
the combined dataframe
}
\description{
Combine data for one simulation, including population metrics, parameter values 
and diversity metrics, and with added columns containing derived variables.
}
\examples{
combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
}
