% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{combine_dfs}
\alias{combine_dfs}
\title{Combine (and optionally process) data for one simulation}
\usage{
combine_dfs(full_dir, include_diversities = TRUE, df_type = "output",
  vaf_cut_off = NA, generation = NA, numcells = NA)
}
\arguments{
\item{full_dir}{base input directory name}

\item{include_diversities}{boolean whether to include diversity metrics (if df_type == "output")}

\item{df_type}{which dataframes to combine}

\item{vaf_cut_off}{exclude genotypes with vaf lower than cut off (if df_type == "genotype_properties" or "driver_genotype_properties")}

\item{generation}{Generation at which to filter (default NA corresponds to no filtering)}

\item{numcells}{Number of cells at which to filter (takes precedent over generation; default NA corresponds to no filtering)}
}
\value{
the combined dataframe
}
\description{
Combine (and optionally process) data for one simulation
}
\examples{
combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "driver_genotype_properties")
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "genotype_properties", vaf_cut_off = 0.002)
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "genotype_counts", numcells = 100)
}
