% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{combine_dfs}
\alias{combine_dfs}
\title{Combine data for one simulation, including population metrics, parameter values 
and (optionally) diversity metrics, and with added columns containing derived variables.}
\usage{
combine_dfs(full_dir, include_diversities = TRUE, df_type = "output",
  vaf_cut_off = NA)
}
\arguments{
\item{full_dir}{base input directory name}

\item{include_diversities}{boolean whether to include diversity metrics}

\item{df_type}{which dataframes to combine}

\item{vaf_cut_off}{exclude genotypes with vaf lower cut off from combined_df}
}
\value{
the combined dataframe
}
\description{
Combine data for one simulation, including population metrics, parameter values 
and (optionally) diversity metrics, and with added columns containing derived variables.
}
\examples{
combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "driver_genotype_properties")
combine_dfs(full_dir = system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
df_type = "genotype_properties", vaf_cut_off = 0.002)
}
