% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{combine_dfs}
\alias{combine_dfs}
\title{Add columns to an output dataframe, containing parameter values and diversity metrics}
\usage{
combine_dfs(full_dir, res = data.frame())
}
\arguments{
\item{full_dir}{base input directory name}

\item{res}{dataframe to which the result will be appended (default is an empty dataframe)}
}
\value{
the same dataframe with additional columns
}
\description{
Add columns to an output dataframe, containing parameter values and diversity metrics
}
\examples{
combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
}
