% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate}
\alias{disp_rate}
\title{Expected disperal speed for migration model}
\usage{
disp_rate(r1, r2, K, m, migration_type = 0, symmetric = FALSE,
  two_dim = TRUE, d = NA)
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{migration_type}{0 or 1 (see details)}

\item{symmetric}{whether migration occurs in both directions}

\item{two_dim}{whether to adjust for two-dimensional growth}

\item{d}{migration distance relative to 1/sqrt(K)}
}
\value{
The growth rate of the radius, measured in cells (not demes).
}
\description{
Expected disperal speed for migration model, allowing dispersal to occur before fixation
}
\examples{
disp_rate(1, 1.1, 2, 0.1, 0)
disp_rate(1, 1.1, 2, 0.1, 1)

# no difference between migration_type = 0 and migration_type = 1 when r1*r2 = 1:
sapply(0:1, disp_rate, r1 = 0.5, r2 = 2, K = 32, m = 0.1)

# otherwise expect a difference:
sapply(0:1, disp_rate, r1 = 0.5, r2 = 3, K = 32, m = 0.1)
}
