% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate}
\alias{disp_rate}
\title{Expected disperal speed, allowing dispersal to occur before fixation}
\usage{
disp_rate(r1, r2, K, d, m, selection = "birth", symmetric = FALSE,
  two_dim = FALSE)
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{selection}{one of "birth", "death", "birth & migration"}

\item{symmetric}{whether migration occurs in both directions}

\item{two_dim}{whether to adjust for two-dimensional growth}
}
\value{
A number, corresponding to the speed.
}
\description{
Expected disperal speed, allowing dispersal to occur before fixation
}
\details{
Adjustment for two-dimensional growth entails multiplying the migration rate by 0.4, which is
a factor estimated from simulation results.
}
\examples{
disp_rate(1, 1.1, 2, 1, 0.1, "birth")
disp_rate(1, 1.1, 2, 1, 0.1, "death")

\dontrun{
df <- combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
num_parameters <- count_parameters(system.file("extdata", "", 
package = "demonanalysis", mustWork = TRUE))
df <- add_columns(df, num_parameters)
median(df$RadiusGrowthRate[which(df$NumCells > 400)]) # dispersal rate from simulation
disp_rate(0, 1, 1, 1, 1, two_dim = TRUE) # predicted dispersal rate
}
}
