% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_demon}
\alias{disp_rate_demon}
\title{A convenient wrapper for calculating dispersal rate for parameter values used in demon.cpp}
\usage{
disp_rate_demon(K, r2, edge_only = 0, two_dim = FALSE)
}
\arguments{
\item{K}{deme carrying capacity}

\item{r2}{birth rate of migrating cells}

\item{edge_only}{whether migration occurs at the edge only}

\item{two_dim}{whether to adjust for two-dimensional growth}
}
\value{
A number, corresponding to the speed.
}
\description{
A convenient wrapper for calculating dispersal rate for parameter values used in demon.cpp
}
\details{
Assumes migration_type = 0, migration_edge_only = 0, migration_rate_scales_with_K = 1.
}
\examples{
disp_rate_demon(8, 1/0.9)
}
