% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_fission}
\alias{disp_rate_fission}
\title{Dispersal speed for deme fission model}
\usage{
disp_rate_fission(r2, K, m, migration_type = 2, migration_edge_only = 0,
  two_dim = TRUE, NumCells = NA)
}
\arguments{
\item{r2}{birth rate relative to reference birth rate}

\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{migration_type}{2 or 3 (see details)}

\item{migration_edge_only}{whether migration occurs at the edge only}

\item{two_dim}{whether to adjust for two-dimensional growth}

\item{NumCells}{total population size (all demes); required only if migration_type = 2 or 3 and migration_edge_only = 0}
}
\value{
The growth rate of the radius, measured in cells (not demes).
}
\description{
Dispersal speed for deme fission model
}
\details{
Assumes that a deme cannot attempt fission until its population size reaches K.

Meaning of migration_type:
\itemize{
  \item 2: fission, such that fission rate is correlated with birth rate, and death rate is uniform
  \item 3: fission, such that fission rate is independent of birth rate, and death rate is uniform
  }
}
\examples{
disp_rate_fission(1, 2, 0.1, NumCells = 1e6)
disp_rate_fission(1, 2, 0.1, migration_type = 3, NumCells = 1e6)

# compare observed and predicted dispersal rates:
df <- combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 
include_diversities = FALSE)
rate_obs <- median(df$RadiusGrowthRate[which(df$NumCells > 400)])
rate_pred <- disp_rate_fission(1, 1, 1, migration_edge_only = 1)
rate_pred / rate_obs # prediction is a bit too low in this case
}
