% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_max}
\alias{disp_rate_max}
\title{Maximum limit on dispersal speed for migration model}
\usage{
disp_rate_max(K, m, r1, r2, migration_type = 0, symmetric = FALSE,
  two_dim = TRUE, d = NA)
}
\arguments{
\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{migration_type}{0 or 1 (see details)}

\item{symmetric}{whether migration occurs in both directions}

\item{two_dim}{whether to adjust for two-dimensional growth}

\item{d}{migration distance relative to 1/sqrt(K); if NA then d is set to sqrt(K)}
}
\value{
The upper limit on the growth rate of the radius, measured in cells (not demes).
}
\description{
Maximum limit on dispersal speed for migration model
}
\details{
Assumes that the time until fixation is negligible relative to the time until migration.

Meaning of migration_type:
\itemize{
  \item 0: migration, such that migration rate is correlated with birth rate, and death rate is uniform
  \item 1: migration, such that migration rate is independent of birth rate, and death rate is uniform
  }
}
\examples{
disp_rate_max(2, 0.1, 1, 1.1, 0)
disp_rate_max(2, 0.1, 1, 1.1, 4)
}
