% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{disp_rate_min}
\alias{disp_rate_min}
\title{Minimum limit on dispersal speed for migration model}
\usage{
disp_rate_min(r1, r2, K, d, m, migration_type = 0, symmetric = FALSE,
  two_dim = TRUE)
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{migration_type}{0, 1 or 4 (see details)}

\item{symmetric}{whether migration occurs in both directions}

\item{two_dim}{whether to adjust for two-dimensional growth}
}
\value{
The dispersal speed.
}
\description{
Minimum limit on dispersal speed for migration model
}
\details{
Assumes that a cell type cannot attempt to migrate until it reaches fixation.

Meaning of migration_type:
0: migration, such that migration rate is correlated with birth rate, and death rate is uniform
1: migration, such that migration rate is independent of birth rate, and death rate is uniform
}
\examples{
disp_rate_min(1, 1.1, 2, 1, 0.1, 0)
disp_rate_min(1, 1.1, 2, 1, 0.1, 4)
}
