% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{dominant}
\alias{dominant}
\title{Find the latest-occurring mutation shared by at least a threshold proportion of cells at a specified time}
\usage{
dominant(anc, pop_df, threshold, generation = NA)
}
\arguments{
\item{anc}{Dataframe of ancestries, as generated by "ancestry" function}

\item{pop_df}{Dataframe with column names including "Identity" and "Population"}

\item{threshold}{Number between 0 and 1; a genotype must exceed this frequency to be considered dominant}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
Identity of the dominant genotype.
}
\description{
Find the latest-occurring mutation shared by at least a threshold proportion of cells at a specified time
}
\examples{
library(ggmuller)
library(dplyr)
phylo <- filter(driver_phylo, CellsPerSample == -1)
pop_df <- get_population_df(phylo)
edges <- get_edges(phylo)
anc <- ancestry(edges)
dominant(anc, pop_df, 0.01)
gens_list <- unique(pop_df$Generation)
dominant(anc, pop_df, 0.01, gens_list[round(length(gens_list) / 2)])
}
