% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{dominant}
\alias{dominant}
\title{Find the latest-occurring mutation shared by at least a threshold proportion of cells}
\usage{
dominant(anc, pop_subdf, threshold)
}
\arguments{
\item{anc}{Dataframe of ancestries, as generated by "ancestry" function}

\item{pop_subdf}{Dataframe with column names "Identity" and "Population"}

\item{threshold}{Number between 0 and 1; a genotype must exceed this frequency to be considered dominant}
}
\description{
Find the latest-occurring mutation shared by at least a threshold proportion of cells
}
\examples{
library(readr)
library(ggmuller)
phylo <- read_delim("data/driver_phylo.dat", "\\t")
phylo <- filter(phylo, CellsPerSample == -1)
pop_df <- get_population_df(phylo)
anc <- ancestry(edges)
pop_subdf <- filter(pop_df, Generation == max(Generation)) \%>\% select(Identity, Population)
dominant(anc, pop_subdf, 0.1)
}
