% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{final_error_message}
\alias{final_error_message}
\title{Return the final line (or an earlier line) of an error log of a simulation}
\usage{
final_error_message(input_dir, adjust = 0)
}
\arguments{
\item{input_dir}{directory name}

\item{adjust}{number of lines prior to the final line (default 0)}
}
\value{
line from error log file
}
\description{
Return the final line (or an earlier line) of an error log of a simulation
}
\examples{
final_error_message(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
final_error_message(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE), 1)
}
