% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{find_correlations}
\alias{find_correlations}
\title{Generic function to find a correlation between two columns of a dataframe}
\usage{
find_correlations(summary, factor1, factor2, result_name, min_count,
  ReturnCI = FALSE, method = "spearman")
}
\arguments{
\item{summary}{dataframe}

\item{factor1}{char name of first column in the summary dataframe}

\item{factor2}{char name of second column in the summary dataframe}

\item{result_name}{char name to give the result}

\item{min_count}{minimum number of items in each column (otherwise result will be NA)}

\item{ReturnCI}{whether to return confidence intervals}
}
\value{
Correlation between the two columns (or NA if either factor1 or factor2 contains NA, 
or if all values of factor1 or factor2 are identical).
Also return the pValue on the spearman correlation coefficient and the associated confidence interval at a 0.95 level.
}
\description{
Generic function to find a correlation between two columns of a dataframe
}
\examples{
s1 <- data.frame(a = 1:3, b = 1:3 * (1 + rnorm(3) / 10))
find_correlations(s1, "a", "b", "c", 3)
# also return 0.95-CI on the spearman correlation coefficient:
find_correlations(s1, "a", "b", "c", 3, TRUE)
}
