% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{first_inc_moment}
\alias{first_inc_moment}
\title{Get the first incomplete moment from frequency data}
\usage{
first_inc_moment(sizes, counts, threshold)
}
\arguments{
\item{sizes}{for example, midpoints of a histogram}

\item{counts}{counts corresponding to the sizes}

\item{threshold}{lower bound of sizes to include}
}
\value{
the first incomplete moment
}
\description{
Get the first incomplete moment from frequency data
}
\examples{
hist1 <- get_genotype_sizes_hist(system.file("extdata", "genotypes.dat", 
package = "demonanalysis", mustWork = TRUE), xmax = 50)
first_inc_moment(hist1$mids, hist1$density, 6)
}
