% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{get_cor_summary}
\alias{get_cor_summary}
\title{Generate summary dataframe of correlations with "outcome"}
\usage{
get_cor_summary(summary, col_names_list, num_parameters = 15)
}
\arguments{
\item{summary}{dataframe including columns named "Generation", "start_time" and "outcome"}

\item{col_names_list}{char vector of column names in the summary dataframe}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}
}
\value{
Dataframe with one row per simulation, and including columns containing the 
correlations between "outcome" and each variable in col_names_list.
}
\description{
Generate summary dataframe of correlations with "outcome"
}
\examples{
get_cor_summary(sum_df, c("DriverDiversity", "DriverEdgeDiversity"))
}
