% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{get_genotype_sizes_hist}
\alias{get_genotype_sizes_hist}
\title{Get a histogram of genotype sizes}
\usage{
get_genotype_sizes_hist(file)
}
\arguments{
\item{file}{name of file containing lists of genotype sizes}
}
\value{
histogram object
}
\description{
Get a histogram of genotype sizes
}
\examples{
hist1 <- get_genotype_sizes_hist("data/genotypes.dat")
}
