% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{get_summary}
\alias{get_summary}
\title{Get summary metrics for one or more simulations}
\usage{
get_summary(data, start_size_range, gap_range, final_size,
  num_parameters = 15)
}
\arguments{
\item{data}{dataframe}

\item{start_size_range}{vector of NumCells at time of initial measurement for forecasting}

\item{gap_range}{vector of projection periods (gap between time of initial measurement and second measurement)}

\item{final_size}{waiting time is measured until tumour reaches this NumCells value}

\item{num_parameters}{number of parameters, accounting for the first set of columns in the dataframe}
}
\value{
a dataframe with one row for each unique combination of parameter values (including seed), 
"gap" and "start_size", (i.e. it summarises over time) and which has added columns "start_time" 
(proportional time until NumCells reached start_size), "end_time" (proportional time until NumCells 
reached end_size), "waiting_time" (difference between start_time and end_time), and "outcome" 
(NumCells after "gap")
}
\description{
Get summary metrics for one or more simulations
}
\examples{
comb_df <- combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
get_summary(comb_df, c(100, 300), c(0.35, 0.65), 1000)
}
