% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{get_summary}
\alias{get_summary}
\title{Get summary metrics for each simulation in a batch}
\usage{
get_summary(data, start_size_range, gap_range, final_size)
}
\arguments{
\item{data}{dataframe}

\item{start_size_range}{vector of Population sizes at time of initial measurement for forecasting}

\item{gap_range}{vector of projection periods (gap between time of initial measurement and second measurement)}

\item{final_size}{waiting time is measured until tumour reaches this Population size}
}
\value{
a dataframe that for each simulation has one row for each combination of "gap" and "start_size", 
and which has added columns "start_time" (proportional time until Population reached start_size), 
"end_time" (proportional time until Population reached end_size), 
"waiting_time" ("start_time" - "end_time"), and "outcome" (Population size after "gap")
}
\description{
Get summary metrics for each simulation in a batch
}
\examples{
comb_df <- combine_dfs(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
get_summary(comb_df, c(10, 20), c(0.1, 0.2), 100)
}
