% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{grid_plot}
\alias{grid_plot}
\title{Plot a grid from a properly formatted data frame.}
\usage{
grid_plot(image_df, palette = NA, discrete = FALSE,
  add_legend = FALSE, legend_title = "")
}
\arguments{
\item{image_df}{data frame formatted by image_df_from_grid_file}

\item{palette}{colour palette (default NA)}

\item{discrete}{whether to use a discrete or continuous colour scale (default FALSE)}

\item{add_legend}{whether to add a legend (default FALSE)}

\item{legend_title}{text for legend title (default none)}
}
\value{
a plot object
}
\description{
Plot a grid from a properly formatted data frame.
}
\examples{
image_df <- image_df_from_grid_file(system.file("extdata", 
"output_passengersgrid.dat", package = "demonanalysis", mustWork = TRUE))
grid_plot(image_df)
}
