% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{hist2}
\alias{hist2}
\title{Alternative to base hist function (using dplyr)}
\usage{
hist2(df, breaks)
}
\arguments{
\item{df}{data frame with Frequency and Count columns}

\item{breaks}{a vector giving the breakpoints between histogram bins}
}
\value{
data frame with densities
}
\description{
Alternative to base hist function (using dplyr)
}
