% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{lambda}
\alias{lambda}
\title{Expected migration rate per deme}
\usage{
lambda(i, m, r1, r2, migration_type = 0, d = NA, K = NA)
}
\arguments{
\item{i}{number of cells in the deme}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{migration_type}{0 or 1 (see details)}

\item{d}{migration distance relative to 1/sqrt(K) (see details)}

\item{K}{deme carrying capacity (only required if d is not NA)}
}
\value{
The migration rate.
}
\description{
Expected migration rate per deme
}
\details{
If migration_type = 0 (meaning that migration is attempted only after a birth event) 
then the migration rate will be multiplied by a factor of r2 * r1. Otherwise no adjustment will be made.

If d and K are set then the calculation assumes that min(d / sqrt(K), 1) is the probability 
that a migration attempt will land outside the deme. 

Importantly, this function does not account for the chance that the migrating cell will 
land in a deme that already occupied.
}
\examples{
# if r1 * r2 = 1 then results are identical for migration_type = 0 and migration_type = 1: 
lambda(1, 0.1, 0.9, 1/0.9, 0)
lambda(1, 0.1, 0.9, 1/0.9, 1)
}
