% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{lambda_invasion}
\alias{lambda_invasion}
\title{Rate of migration followed by survival}
\usage{
lambda_invasion(i, K, d, m, r1, r2, migration_type = 0, death_rate = 1)
}
\arguments{
\item{i}{number of cells of the focus type}

\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{migration_type}{0 or 1 (see details)}

\item{death_rate}{death rate of migrating cells, relative to death rate of cells in destination deme}
}
\value{
The rate of migration followed by survival.
}
\description{
Rate of migration followed by survival
}
\details{
If migration_type = 0 (meaning that migration rate is correlated with birth rate) 
then the migration rate will be multiplied by a factor of r2 * r1. Otherwise no adjustment will be made.
}
\examples{
lambda_invasion(1, 2, 1, 0.1, 1, 1.1, 0)
lambda_invasion(1, 2, 1, 0.1, 1, 1.1, 1)
}
