% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacollation.R
\name{parameter_names_and_values}
\alias{parameter_names_and_values}
\title{Find the names and final values of parameters that were varied in a batch of simulations}
\usage{
parameter_names_and_values(input_dir)
}
\arguments{
\item{input_dir}{base input directory name}
}
\value{
dataframe of parameter names and values, or NA if input_dir contains results of a single simulation
}
\description{
Find the names and final values of parameters that were varied in a batch of simulations
}
\examples{
parameter_names_and_values(system.file("example_batch", "", 
package = "demonanalysis", mustWork = TRUE))
}
