% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_all_charts}
\alias{plot_all_charts}
\title{Plot a set of charts representing allele frequencies and genotype sizes}
\usage{
plot_all_charts(path_or_dflist, output_filename = NA, file_type = "png",
  output_dir = NA, max_size = NA, generation = NA, numcells = NA,
  counts_max = 10)
}
\arguments{
\item{path_or_dflist}{folder containing the input files, or a list of data frames}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}

\item{max_size}{maximum size (default NA corresponds to plotting frequencies, not sizes)}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}

\item{numcells}{Number of cells at which to make the measurement (takes precedent over generation; default NA corresponds to the final size)}

\item{counts_max}{Max value of y-axis in counts plot}
}
\value{
plot displyed on screen
}
\description{
Plot a set of charts representing allele frequencies and genotype sizes
}
\examples{
plot_all_charts(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
plot_all_charts(list(output_allele_counts, output_driver_allele_counts, 
output_genotype_counts, output_driver_genotype_counts))
}
