% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_all_images}
\alias{plot_all_images}
\title{Create a set of grid and Muller plots}
\usage{
plot_all_images(path, output_filename = NA, file_type = "png",
  output_dir = NA, trim = -1)
}
\arguments{
\item{path}{folder containing the input files}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}

\item{trim}{how many rows and columns to remove from grids; if trim < 0 (default) then all rows and columns containing NA are removed}
}
\value{
either an image file or a plot displyed on screen
}
\description{
Create a set of grid and Muller plots
}
\examples{
plot_all_images(system.file("extdata", "", package = "demonanalysis", mustWork = TRUE))
}
