% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_allele_cum_dist}
\alias{plot_allele_cum_dist}
\title{Plot cumulative density of variant allele frequencies versus inverse allele frequency}
\usage{
plot_allele_cum_dist(file, generation = NA)
}
\arguments{
\item{file}{file containing columns "Frequency" and "Count"}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
plot displyed on screen
}
\description{
Plot cumulative density of variant allele frequencies versus inverse allele frequency
}
\examples{
plot_allele_cum_dist(system.file("extdata", "output_allele_counts.dat", 
package = "demonanalysis", mustWork = TRUE))
}
