% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_allele_cum_dist}
\alias{plot_allele_cum_dist}
\title{Plot cumulative density of variant allele frequencies versus inverse allele frequency}
\usage{
plot_allele_cum_dist(df, generation = NA)
}
\arguments{
\item{df}{data frame containing "inverse_frequency" and "cumulative_count" columns}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
plot displyed on screen
}
\description{
Plot cumulative density of variant allele frequencies versus inverse allele frequency
}
\examples{
plot_allele_cum_dist(output_allele_cum_dist)
}
