% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_allele_cum_dist}
\alias{plot_allele_cum_dist}
\title{Plot cumulative density of variant allele frequencies versus inverse allele frequency}
\usage{
plot_allele_cum_dist(df)
}
\arguments{
\item{df}{data frame containing "inverse_frequency" and "cumulative_count" columns}
}
\value{
plot displyed on screen
}
\description{
Plot cumulative density of variant allele frequencies versus inverse allele frequency
}
\examples{
library(readr)
output_allele_cum_dist <- read_delim("data/output_allele_cum_dist.dat", "\\t", trim_ws = TRUE)
plot_allele_cum_dist(output_allele_cum_dist)
}
