% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_allele_hist}
\alias{plot_allele_hist}
\title{Plot a histogram of variant allele frequencies}
\usage{
plot_allele_hist(df, generation = NA)
}
\arguments{
\item{df}{data frame containing "frequency" and "density" columns}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
plot displyed on screen
}
\description{
Plot a histogram of variant allele frequencies
}
\examples{
plot_allele_hist(output_allele_hist)
}
