% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_allelecount_vs_origintime}
\alias{plot_allelecount_vs_origintime}
\title{Plot allele count versus origin time, coloured by birth rate}
\usage{
plot_allelecount_vs_origintime(file, log = FALSE)
}
\arguments{
\item{file}{file containing columns "AlleleCount", "OriginTime and "BirthRate"}

\item{log}{if TRUE then y-axis will be log-transformed (default FALSE)}
}
\value{
plot displyed on screen
}
\description{
Plot allele count versus origin time, coloured by birth rate
}
\examples{
plot_allelecount_vs_origintime(system.file("extdata", "output_genotype_properties.dat", 
package = "demonanalysis", mustWork = TRUE))
}
