% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_common_allele_counts}
\alias{plot_common_allele_counts}
\title{Plot counts of variant allele frequencies for the most common alleles}
\usage{
plot_common_allele_counts(df)
}
\arguments{
\item{df}{dataframe with columns named "allele_frequency" and "count"}
}
\value{
plot
}
\description{
Plot counts of variant allele frequencies for the most common alleles
}
\examples{
df1 <- get_common_allele_counts(system.file("extdata", "output_allele_freqs.dat", 
package = "demonanalysis", mustWork = TRUE))
plot_common_allele_counts(df1)
}
