% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_corr_waiting_time_versus_depth}
\alias{plot_corr_waiting_time_versus_depth}
\title{Plot correlations between waiting time and diversity measured at different depths, coloured by size at measurement}
\usage{
plot_corr_waiting_time_versus_depth(df, col_prefix, output_filename = NA,
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{dataframe generated by get_wait_cor_summary}

\item{col_prefix}{text that appears in all names of columns containing correlation statistics}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Plot correlations between waiting time and diversity measured at different depths, coloured by size at measurement
}
\examples{
library(dplyr)
plot_corr_waiting_time_versus_depth(filter(depth_wait_cor_summary, K == 16), 
"DriverDiversityFrom1SamplesAtDepth")
}
