% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_driver_genotype_freq_hist}
\alias{plot_driver_genotype_freq_hist}
\title{Plot a histogram of driver genotype frequencies for the most common driver genotypes}
\usage{
plot_driver_genotype_freq_hist(file, generation = NA)
}
\arguments{
\item{file}{name of "phylo" file}

\item{generation}{Generation at which to make the measurement (default NA corresponds to the final Generation)}
}
\value{
plot displyed on screen
}
\description{
Plot a histogram of driver genotype frequencies for the most common driver genotypes
}
\examples{
plot_driver_genotype_freq_hist(system.file("extdata", "driver_phylo.dat", 
package = "demonanalysis", mustWork = TRUE))
}
