% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_first_inc_moment}
\alias{plot_first_inc_moment}
\title{Plot first incomplete moment from frequency data}
\usage{
plot_first_inc_moment(sizes, counts, ...)
}
\arguments{
\item{sizes}{for example, midpoints of a histogram}

\item{counts}{counts corresponding to the sizes}

\item{...}{other parameters passed to plot}
}
\value{
plot displyed on screen
}
\description{
Plot first incomplete moment from frequency data
}
\examples{
df_test <- data.frame(size = 1:20, count = exp(-(1:20)))
plot_first_inc_moment(df_test$size, df_test$count)
}
