% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_first_inc_moment}
\alias{plot_first_inc_moment}
\title{Plot first incomplete moment of genotype sizes}
\usage{
plot_first_inc_moment(hist)
}
\arguments{
\item{hist}{histogram of genotype sizes}
}
\value{
plot displyed on screen
}
\description{
Plot first incomplete moment of genotype sizes
}
\examples{
hist1 <- get_genotype_sizes_hist(system.file("extdata", "genotypes.dat", package = "demonanalysis"))
plot_first_inc_moment(hist1)
}
