% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demonanalysis.R
\name{plot_genotype_sizes_hist}
\alias{plot_genotype_sizes_hist}
\title{Plot a histogram of genotype sizes}
\usage{
plot_genotype_sizes_hist(hist)
}
\arguments{
\item{hist}{histogram of genotype sizes}
}
\value{
plot displyed on screen
}
\description{
Plot a histogram of genotype sizes
}
\examples{
hist1 <- get_genotype_sizes_hist(system.file("extdata", "genotypes.dat", 
package = "demonanalysis", mustWork = TRUE))
plot_genotype_sizes_hist(hist1)
}
