% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastingplots.R
\name{plot_viol_sweep_metric}
\alias{plot_viol_sweep_metric}
\title{Violin plot of mean sweep metric (genotype frequency autocorrelation)}
\usage{
plot_viol_sweep_metric(df, output_filename = "viol_sweep_metric",
  file_type = "png", output_dir = NA)
}
\arguments{
\item{df}{dataframe generated by get_summary (filtered), get_cor_summary (filtered) or get_wait_cor_summary (filtered)}

\item{output_filename}{name of output image file}

\item{file_type}{either "pdf" or "png" (other values default to "pdf")}

\item{output_dir}{folder in which to save the image file; if NA then plots are displayed on screen instead}
}
\value{
a plot object
}
\description{
Violin plot of mean sweep metric (genotype frequency autocorrelation)
}
\examples{
library(dplyr)
plot_viol_sweep_metric(filter(summary, gap == min(gap), start_size == min(start_size)))
plot_viol_sweep_metric(filter(cor_summary, gap == min(gap), start_size == min(start_size)))
plot_viol_sweep_metric(filter(wait_cor_summary, start_size == min(start_size)))
}
