% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{quadratic_diversity}
\alias{quadratic_diversity}
\title{Rao's quadratic diversity, using a binary distance measure, converted to the effective number of species}
\usage{
quadratic_diversity(df, sigma, threshold = 0.1)
}
\arguments{
\item{df}{Dataframe with two columns containing trait values and their frequencies or abundancies}

\item{sigma}{Cutoff for considering two trait values to be identical}

\item{threshold}{minimum trait value to include in calculation}
}
\value{
effective number of species
}
\description{
Rao's quadratic diversity, using a binary distance measure, converted to the effective number of species
}
\examples{
df <- data.frame(value = 1:5, count = 1:5)
quadratic_diversity(df, 1)
}
