% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{sweep_sequence}
\alias{sweep_sequence}
\title{Calculate how the genotype frequency distribution changes over time}
\usage{
sweep_sequence(pop_df, lag_type = "generations", breaks = 10,
  lag_gens = 500)
}
\arguments{
\item{pop_df}{Dataframe with column names "Identity", "Population" and "Generation"}

\item{lag_type}{Either "generations" or "proportions" (default "generations")}

\item{breaks}{Number of breaks for determining lag (used only if lag_type = "proportions"; default 10)}

\item{lag_gens}{Lag in terms of generations (used only if lag_type = "generations"; default 500)}
}
\value{
For each generation g in pop_df, excluding the first generation, the output vector quantifies 
the change in genotype frequencies compared to generation g - lag_gens (by summing the squares of the 
differences). The length of the output sequence is the same as the number of rows in the input dataframe.
The first value is always zero.
}
\description{
Calculate how the genotype frequency distribution changes over time
}
\examples{
library(ggmuller)
library(dplyr)
phylo <- filter(driver_phylo, CellsPerSample == -1)
pop_df <- get_population_df(phylo)
sweep_seq1 <- sweep_sequence(pop_df, lag_type = "proportions", breaks = 6)
lag_gens <- round(length(unique(pop_df$Generation))/6)
sweep_seq2 <- sweep_sequence(pop_df, lag_type = "generations", lag_gens = lag_gens)
identical(sweep_seq1, sweep_seq2)
sweep_df <- data.frame(y = sweep_seq2, x = (1:length(sweep_seq2))/length(sweep_seq2))
plot(y ~ x , data = sweep_df, type = "l")
}
