% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweepmetrics.R
\name{sweep_times}
\alias{sweep_times}
\title{Find generations at which sweeps (or at least partial sweeps) occur}
\usage{
sweep_times(phylo, threshold, min_pop = 0)
}
\arguments{
\item{phylo}{Dataframe containing phylogeny}

\item{threshold}{Number between 0 and 1; a genotype must exceed this frequency to be considered dominant}

\item{min_pop}{Genotypes with populations smaller than this size will be removed before analysis (default 0)}
}
\description{
Find generations at which sweeps (or at least partial sweeps) occur
}
\examples{
library(readr)
phylo <- read_delim("data/driver_phylo.dat", "\\t")
phylo <- filter(phylo, CellsPerSample == -1)
sweep_times(phylo, threshold = 0.9)
}
