% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{T_grow_j}
\alias{T_grow_j}
\title{Mean time from population size 1 until size j <= K in a Moran process}
\usage{
T_grow_j(j, r1, r2, K)
}
\arguments{
\item{j}{final population size of the focus type}

\item{r1}{birth rate of non-focus cells, relative to reference birth rate}

\item{r2}{birth rate of focus cell type}

\item{K}{deme carrying capacity}
}
\value{
A number, corresponding to the time.
}
\description{
Mean time from population size 1 until size j <= K in a Moran process
}
\details{
See Traulsen 2009, equations 1.38 and 1.39.
}
\examples{
T_grow_j(3, 1, 1.1, 4)
}
