% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{time_expected}
\alias{time_expected}
\title{Migration waiting time (any population size)}
\usage{
time_expected(r1, r2, K, m, migration_type = 0, d = NA)
}
\arguments{
\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{K}{deme carrying capacity}

\item{m}{migration rate per cell, relative to birth rate}

\item{migration_type}{0 or 1 (see details)}

\item{d}{migration distance relative to 1/sqrt(K); if NA then d is set to sqrt(K)}
}
\value{
The waiting time.
}
\description{
Expected time of first successful migration, which may occur before fixation
}
\details{
Meaning of migration_type:
0: migration, such that migration rate is correlated with birth rate, and death rate is uniform
1: migration, such that migration rate is independent of birth rate, and death rate is uniform
}
\examples{
time_expected(1, 1.1, 2, 0.1, 0)
time_expected(1, 1.1, 2, 0.1, 4)

# for comparison:
time_migration(2, 0.1, 1, 1.1, 0)
time_migration(2, 0.1, 1, 1.1, 4)
}
