% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speedestimates.R
\name{time_migration}
\alias{time_migration}
\title{Mean time until successful migration from a deme that is fully occupied by the focus cell type}
\usage{
time_migration(K, d, m, r1, r2, selection = "birth")
}
\arguments{
\item{K}{deme carrying capacity}

\item{d}{migration distance relative to 1/sqrt(K)}

\item{m}{migration rate per cell, relative to birth rate}

\item{r1}{birth rate of cells in the destination deme, relative to reference birth rate}

\item{r2}{birth rate of migrating cells, relative to r1}

\item{selection}{one of "birth", "death", "birth & migration"}
}
\value{
A number, corresponding to the time.
}
\description{
Mean time until successful migration from a deme that is fully occupied by the focus cell type
}
\examples{
time_migration(2, 1, 0.1, 1, 1.1, "birth")
time_migration(2, 1, 0.1, 1, 1.1, "death")
}
